Change minimum SDK version
Update the minimum SDK version required for the application. This with the
goal of reaching more people. After some research it turned out that (in
constrast to my original believes) there are still quite a few phones out
that that still run Android 4.4 KitKat (which uses SDK 19 as opposed to
21). The number of Android phones is ~19% while Lollipop (5.0 and 5.1
combined) is ~32% and Marshmallow (6.0) ~31%.

Source: https://developer.android.com/about/dashboards/index.html