Let keys write themselves as JSON string
Previously the GsonKeyAdapter class decided how keys are encoded to JSON
but putting this code in the Key classes itself (enforced via the IKey
interface) makes more sense. It provides the code with direct access to
the attributes of the key and ensures that each key can be transformed
into a JSON string.