Let keys read themselves as JSON strings
Keys now have a static method which can be used to create a new instance
of that key given the JsonReader. However this is not the prettiest
solution since the static method cannot be enforced by the IKey interface.

Also changed how keys are written in JSON, they're now written as a array
with two elements. The first of which is a string of the type and the
second of which is the object created by the `.writeJSON()` method.

Resolves #67