Distribute androidFilesDir via ServiceLocator
Updated the ServiceLocator (and its interface) to allow for the
distribution of the androidFilesDir (i.e. the result of `getFilesDir()` in
an activity. This also fixes the (reintroduced) issue that the database
cannot instantiate the JSON files.

Also added missing `@Override`s in the ServiceLocator and did some general
reformatting of the classes already touched.