Map constructors & Static helper methods
Added constructors to the IKey implementation classes which can be used by
the GsonKeyAdapter to create a new IKey of that instance given a map of
the JSON object representing the key.

Also added static helper methods to the RSAKey class which can later be
used to obtain the modulus/exponent of the key without having to edit the
IKey interface.